+++ esp-idf-svc/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * SPI Ethernet (async; a blocking example for RMII Ethernet already exists)
   * TCP client and server (blocking and async)
   * SNTP service
+  * Websocket client (blocking)
 * Breaking changes in module `eventloop`: 
   * Async send and receive functionality now implemented directly on the `esp-idf-svc` event loop types, as the `embedded_svc::utils::asyncify` module is now gone
   * Types `EspTypedEventLoop` and `EspPostbox` are now retired. Use `EspEventLoop` directly, as it has the same functionality